fix: cookies get returned back to their original state
If a user is debugging with hoppscotch using a cookie for authorization, and they have one with the same name set it would just disappear. This commit fixes that bug by trying to return the cookie back to their original state. It also passes `credentials: "omit"` to the actual fetch to prevent it from doing any of the cookie handling as well.